new file with tests
mama
